Code/requirement.txt
autopep8==1.4.3
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql-connector==2.2.9
pkg-resources==0.0.0
pycodestyle==2.5.0
six==1.12.0
Werkzeug==0.15.2
web3==4.9.1
mysql-connector-python==8.0.15
